requirements.txt
ago==0.0.93
aiofiles==0.5.0
aiomysql==0.0.20
aiosqlite==0.15.0
amqp==2.5.2
ansicon==1.89.0
appdirs==1.4.3
arrow==0.15.6
asgiref==3.2.7
asyncpg==0.21.0
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
bcrypt==3.2.0
billiard==3.6.3.0
blessed==1.17.5
callofduty.py==1.2.1
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
distlib==0.3.0
django-compat==1.0.15
django-picklefield==2.1.1
ecdsa==0.14.1
fastapi==0.60.1
fastapi-utils==0.2.1
filelock==3.0.12
gevent==20.6.2
greenlet==0.4.16
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.8.25
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
install==1.3.3
iso8601==0.1.12
Jinja2==2.11.2
jinxed==1.0.0
jsonfield==3.1.0
kombu==4.6.8
limits==1.5.1
loguru==0.5.1
MarkupSafe==1.1.1
orjson==3.0.1
passlib==1.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.6.1
PyHamcrest==2.0.2
PyMySQL==0.9.2
pyOpenSSL==19.1.0
PyPika==0.37.16
python-dateutil==2.8.1
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.1
redis==3.5.2
requests==2.23.0
requests-futures==1.0.0
rfc3986==1.4.0
rsa==4.6
service-identity==18.1.0
six==1.15.0
slowapi==0.1.1
sniffio==1.1.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
starlette==0.13.6
tortoise-orm==0.16.14
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
uvicorn==0.11.8
vine==1.3.0
virtualenv==20.0.20
waitress==1.4.3
wcwidth==0.1.9
websockets==8.1
win32-setctime==1.0.1
zope.event==4.4
zope.interface==5.1.0
gunicorn==20.0.4